Fix sheet reopening, allow language switching in disclaimer
