+++ packages/app/src/components/UIOverlay/ReportDetailsSheet.tsx
@@ -23,7 +23,7 @@ export const ReportDetailsSheet = () => {
     if (report === null) return null
 
     return (
-        <FFScrollSheet ref={sheetRef} backdropType="transparent">
+        <FFScrollSheet ref={sheetRef} backdropType="none">
             <ReportItem report={report} />
             {statistics.data !== undefined && statistics.data.numberOfReports > 1 && (
                 <FFText fontSize={15} mt="xs" color="darkText">
